+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add option to set property namespace used to determine a nodes topology.
+
 ## [0.20.0] - 2022-07-20
 
 ### Added